iot/requirements.txt
firebase-admin
pylint
aiohttp
RPi.GPIO
picamera